+++ common/src/main/java/com/cobblemon/mod/common/mixin/client/PlayerModelMixin.java
@@ -7,8 +7,7 @@
  */
 
 package com.cobblemon.mod.common.mixin.client;
-import com.cobblemon.mod.common.api.riding.Rideable;
-import com.cobblemon.mod.common.api.riding.RidingManager;
+
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import net.minecraft.client.model.PlayerModel;
 import net.minecraft.world.entity.Entity;
@@ -22,13 +21,16 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(PlayerModel.class)
 public class PlayerModelMixin {
     @Inject(method = "setupAnim(Lnet/minecraft/world/entity/LivingEntity;FFFFF)V", at = @At(value = "HEAD"))
-    private void cobblemon$setHeadRotation(LivingEntity entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch, CallbackInfo ci){
+    private void cobblemon$setHeadRotation(LivingEntity entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch, CallbackInfo ci) {
         if (!(entity instanceof Player)) return;
         Entity vehicle = entity.getVehicle();
 
-        if (!(vehicle instanceof PokemonEntity)) return;
-        RidingManager ridingManager = ((Rideable) vehicle).getRiding();
-        if (ridingManager.shouldRotatePlayerHead((PokemonEntity) vehicle)) return;
+        if (!(vehicle instanceof PokemonEntity pokemonEntity)) return;
+
+        var shouldRotatePlayerHead = pokemonEntity.ifRidingAvailableSupply(false, (behaviour, settings, state) -> {
+            return behaviour.shouldRotatePlayerHead(settings, state, pokemonEntity);
+        });
+        if (shouldRotatePlayerHead) return;
 
         netHeadYaw = 0f;
         headPitch = 0f;