+++ common/src/main/kotlin/com/cobblemon/mod/common/client/entity/PokemonClientDelegate.kt
@@ -12,6 +12,7 @@ import com.bedrockk.molang.runtime.struct.QueryStruct
 import com.bedrockk.molang.runtime.value.DoubleValue
 import com.bedrockk.molang.runtime.value.StringValue
 import com.cobblemon.mod.common.Cobblemon
+import com.cobblemon.mod.common.CobblemonNetwork
 import com.cobblemon.mod.common.CobblemonSounds
 import com.cobblemon.mod.common.api.entity.PokemonSideDelegate
 import com.cobblemon.mod.common.api.molang.MoLangFunctions.addFunctions
@@ -31,6 +32,7 @@ import com.cobblemon.mod.common.client.render.models.blockbench.animation.Primar
 import com.cobblemon.mod.common.client.render.models.blockbench.repository.VaryingModelRepository
 import com.cobblemon.mod.common.client.render.pokemon.PokemonRenderer.Companion.ease
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity
+import com.cobblemon.mod.common.net.messages.server.pokemon.update.ServerboundUpdateRidingStatePacket
 import com.cobblemon.mod.common.pokemon.Pokemon
 import com.cobblemon.mod.common.util.MovingSoundInstance
 import com.cobblemon.mod.common.util.asExpressionLike
@@ -103,6 +105,7 @@ class PokemonClientDelegate : PosableState(), PokemonSideDelegate {
                 currentEntity.pokemon.species = PokemonSpecies.getByIdentifier(identifier)!! // TODO exception handling
                 // force a model update - handles edge case where the PosableState's tracked PosableModel isn't updated until the LivingEntityRenderer render is run
                 currentModel = VaryingModelRepository.getPoser(identifier, this)
+                currentEntity.refreshRiding()
             } else if (data == PokemonEntity.ASPECTS) {
                 currentAspects = currentEntity.entityData.get(PokemonEntity.ASPECTS)
                 currentEntity.pokemon.shiny = currentAspects.contains("shiny")
@@ -376,6 +379,18 @@ class PokemonClientDelegate : PosableState(), PokemonSideDelegate {
     override fun tick(entity: PokemonEntity) {
         incrementAge(entity)
         playWildShinySounds()
+        sendRidingChanges(entity)
+    }
+
+    private fun sendRidingChanges(entity: PokemonEntity) {
+        val player = Minecraft.getInstance().player ?: return
+        if (entity.controllingPassenger != player) return
+        entity.ifRidingAvailable { behaviour, _, state ->
+            if (entity.previousRidingState != null && !state.shouldSync(entity.previousRidingState!!)) {
+                return@ifRidingAvailable
+            }
+            CobblemonNetwork.sendToServer(ServerboundUpdateRidingStatePacket(entity.id, behaviour.key, state))
+        }
     }
 
     fun playWildShinySounds() {