+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/PosableModel.kt
@@ -11,13 +11,12 @@ package com.cobblemon.mod.common.client.render.models.blockbench
 import com.bedrockk.molang.runtime.MoLangRuntime
 import com.bedrockk.molang.runtime.struct.QueryStruct
 import com.cobblemon.mod.common.Cobblemon
-import com.cobblemon.mod.common.Rollable
+import com.cobblemon.mod.common.OrientationControllable
 import com.cobblemon.mod.common.api.molang.ExpressionLike
 import com.cobblemon.mod.common.api.molang.MoLangFunctions.addFunctions
 import com.cobblemon.mod.common.api.molang.MoLangFunctions.setup
 import com.cobblemon.mod.common.client.ClientMoLangFunctions.animationFunctions
 import com.cobblemon.mod.common.client.ClientMoLangFunctions.setupClient
-import com.cobblemon.mod.common.client.entity.NPCClientDelegate
 import com.cobblemon.mod.common.client.entity.PokemonClientDelegate
 import com.cobblemon.mod.common.client.render.AnimatedModelTextureSupplier
 import com.cobblemon.mod.common.client.render.ModelLayer
@@ -41,6 +40,7 @@ import com.cobblemon.mod.common.entity.npc.NPCEntity
 import com.cobblemon.mod.common.entity.pokeball.EmptyPokeBallEntity
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity
 import com.cobblemon.mod.common.util.asExpressionLike
+import com.cobblemon.mod.common.util.cobblemonResource
 import com.cobblemon.mod.common.util.plus
 import com.cobblemon.mod.common.util.toRGBA
 import com.mojang.blaze3d.systems.RenderSystem
@@ -49,7 +49,9 @@ import com.mojang.blaze3d.vertex.PoseStack
 import com.mojang.blaze3d.vertex.VertexConsumer
 import com.mojang.blaze3d.vertex.VertexFormat
 import com.mojang.math.Axis
+import net.minecraft.client.Minecraft
 import net.minecraft.client.model.geom.ModelPart
+import net.minecraft.client.player.RemotePlayer
 import net.minecraft.client.renderer.MultiBufferSource
 import net.minecraft.client.renderer.RenderStateShard
 import net.minecraft.client.renderer.RenderType
@@ -602,13 +604,15 @@ open class PosableModel(@Transient override val rootPart: Bone) : ModelFrame {
 
         val primaryAnimation = state.primaryAnimation
         val shouldRotateHead = if (entity is PokemonEntity) {
-            entity.riding.shouldRotatePokemonHead(entity)
-        } else {
-            true
-        }
+            entity.ifRidingAvailableSupply(true) { behaviour, settings, ridingState ->
+                behaviour.shouldRotatePokemonHead(settings, ridingState, entity)
+            }
+        } else true
+
 
         val headPitch = if (shouldRotateHead) headPitch else 0f
         val headYaw = if (shouldRotateHead) headYaw else 0f
+
 
         // Quirks will run if there is no primary animation running and quirks are enabled for this context.
         if (primaryAnimation == null && context.request(RenderContext.DO_QUIRKS) != false) {
@@ -722,16 +726,24 @@ open class PosableModel(@Transient override val rootPart: Bone) : ModelFrame {
         var scale = 1F
         // We could improve this to be generalized for other entities. First we'd have to figure out wtf is going on, though.
         if (entity is PokemonEntity) {
-            if(entity.passengers.isNotEmpty() && entity.controllingPassenger is Rollable && (entity.controllingPassenger as Rollable).orientation != null){
-                val controllingPassenger = entity.controllingPassenger
+            if(entity.passengers.isNotEmpty() && entity.controllingPassenger is OrientationControllable
+                && (entity.controllingPassenger as OrientationControllable).orientationController.active){
+                val controllingPassenger = entity.controllingPassenger as OrientationControllable
+                val controller = controllingPassenger.orientationController
                 val transformationMatrix = Matrix4f()
                 val center = Vector3f(0f, entity.bbHeight/2, 0f)
                 transformationMatrix.translate(center)
-                transformationMatrix.mul(Matrix4f((controllingPassenger as Rollable).orientation!!))
+                transformationMatrix.rotate(controller.getRenderOrientation(state.getPartialTicks()))
                 transformationMatrix.translate(center.negate())
                 matrixStack.mulPose(transformationMatrix)
+            } else if( entity.passengers.isNotEmpty()) {
+                // rotate a ridden but non rollable pokemon correctly
+                val driver = entity.controllingPassenger ?: entity
+                var yRot = Mth.lerp(state.getPartialTicks(), driver.yRotO, driver.yRot)
+                yRot = Mth.wrapDegrees(yRot)
+                matrixStack.mulPose(Axis.YP.rotationDegrees(180 - yRot))
             } else {
-                var yRot = Mth.lerp(state.getPartialTicks(), entity.yBodyRotO, entity.yBodyRot)
+                var yRot = Mth.lerp(state.getPartialTicks(), entity.yRotO, entity.yRot)
                 yRot = Mth.wrapDegrees(yRot)
                 matrixStack.mulPose(Axis.YP.rotationDegrees(180 - yRot))
             }