+++ common/src/main/kotlin/com/cobblemon/mod/common/pokemon/Pokemon.kt
@@ -23,6 +23,8 @@ import com.cobblemon.mod.common.api.events.CobblemonEvents.FRIENDSHIP_UPDATED
 import com.cobblemon.mod.common.api.events.CobblemonEvents.POKEMON_FAINTED
 import com.cobblemon.mod.common.api.events.pokemon.*
 import com.cobblemon.mod.common.api.events.pokemon.healing.PokemonHealedEvent
+import com.cobblemon.mod.common.api.mark.Mark
+import com.cobblemon.mod.common.api.mark.Marks
 import com.cobblemon.mod.common.api.molang.MoLangFunctions.addPokemonFunctions
 import com.cobblemon.mod.common.api.molang.ObjectValue
 import com.cobblemon.mod.common.api.moves.BenchedMove
@@ -245,7 +247,11 @@ open class Pokemon : ShowdownIdentifiable {
             onChange(NicknameUpdatePacket({ this }, value))
         }
 
-    fun getDisplayName(): MutableComponent = nickname?.copy()?.takeIf { it.contents != PlainTextContents.EMPTY } ?: species.translatedName.copy()
+    fun getDisplayName(showTitle: Boolean = false): MutableComponent {
+        var name = nickname?.copy()?.takeIf { it.contents != PlainTextContents.EMPTY } ?: species.translatedName.copy()
+        if (showTitle) activeMark?.let { name = it.getTitle(name) }
+        return name
+    }
 
     var level = 1
         set(value) {
@@ -505,6 +511,30 @@ open class Pokemon : ShowdownIdentifiable {
             onChange(CosmeticItemUpdatePacket({ this }, value))
         }
 
+    var activeMark: Mark? = null
+        set(value) {
+            field = value
+            onChange(ActiveMarkUpdatePacket({ this }, value))
+        }
+
+    var marks: MutableSet<Mark> = mutableSetOf()
+        set(value) {
+            field = value
+            onChange(MarksUpdatePacket({ this }, value))
+        }
+
+    var potentialMarks: MutableSet<Mark> = mutableSetOf()
+        set(value) {
+            field = value
+            onChange(MarksPotentialUpdatePacket({ this }, value))
+        }
+
+    var markings: List<Int> = listOf(0, 0, 0, 0, 0, 0)
+        set(value) {
+            field = value
+            onChange(MarkingsUpdatePacket({ this }, value))
+        }
+
     fun asRenderablePokemon() = RenderablePokemon(species, aspects)
 
     /**
@@ -589,7 +619,6 @@ open class Pokemon : ShowdownIdentifiable {
         }
     }
 
-
     open fun getStat(stat: Stat) = Cobblemon.statProvider.getStatForPokemon(this, stat)
 
     /**
@@ -633,7 +662,6 @@ open class Pokemon : ShowdownIdentifiable {
         illusion: IllusionEffect? = null,
         mutation: (PokemonEntity) -> Unit = {},
     ): CompletableFuture<PokemonEntity> {
-
 
         // Handle special case of shouldered Cobblemon
         if (this.state is ShoulderedState) {
@@ -1008,6 +1036,57 @@ open class Pokemon : ShowdownIdentifiable {
      */
     fun removeHeldItem(): ItemStack = this.swapHeldItem(ItemStack.EMPTY)
 
+    fun addPotentialMark(mark: Mark) {
+        potentialMarks.add(mark)
+        if (!isClient) onChange(MarkPotentialAddUpdatePacket({ this }, mark))
+    }
+
+    fun exchangeMark(mark: Mark, give: Boolean) {
+        if (give) {
+            mark.replace?.takeIf { it.isNotEmpty() }?.let { replacements ->
+                for (mark in replacements) Marks.getByIdentifier(mark)?.let { exchangeMark(it, false) }
+            }
+
+            marks.add(mark)
+            if (!isClient) onChange(MarkAddUpdatePacket({ this }, mark))
+        }
+        else {
+            marks.remove(mark)
+            if (activeMark == mark) activeMark = null
+            if (!isClient) onChange(MarkRemoveUpdatePacket({ this }, mark))
+        }
+    }
+
+    /**
+     * Calculate mark to give from list of potential marks.
+     * @return True if a mark has been applied, false otherwise
+     */
+    fun applyPotentialMarks(): Boolean {
+        // Remove any marks that are already owned
+        val potentials = potentialMarks.filterNot { mark -> marks.contains(mark) }.toMutableSet()
+
+        if (!potentials.isEmpty()) {
+            var selectedMark: Mark? = null
+            val sortedMarkGroups = potentials.groupBy { it.getChanceGroup() }.toSortedMap(compareBy { it.second })
+
+            for ((chanceGroup, group) in sortedMarkGroups) {
+                val probability = chanceGroup.second.coerceIn(0F, 1F) * 100
+                val randomValue = Random.nextDouble(0.0, 100.0)
+
+                if (randomValue < probability) {
+                    selectedMark = group.random()
+                    break
+                }
+            }
+            potentialMarks = mutableSetOf()
+            selectedMark?.let {
+                exchangeMark(it, true)
+                return true
+            }
+        }
+        return false
+    }
+
     fun saveToNBT(registryAccess: RegistryAccess, nbt: CompoundTag = CompoundTag()): CompoundTag {
         return this.saveTo(registryAccess.createSerializationContext(NbtOps.INSTANCE), nbt).orThrow as CompoundTag
     }
@@ -1108,6 +1187,12 @@ open class Pokemon : ShowdownIdentifiable {
         this.forcedAspects = other.forcedAspects
         this.features = other.features
         this.cosmeticItem = other.cosmeticItem
+        this.activeMark = other.activeMark
+        this.marks.clear()
+        this.marks += other.marks
+        this.potentialMarks.clear()
+        this.potentialMarks += other.marks
+        this.markings = other.markings
         this.updateAspects()
         this.refreshOriginalTrainer()
         this.initialize()